Add the `PseudoPotentialData` base class data plugin
This data plugin forms a base class for other plugins that will
represent a pseudo potential. The assumptions are that each pseudo
potential will always consist of a single file, which is why it
subclasses the `SingleFileData` data plugin. The md5 is of the file is
automatically stored as an attribute.

Additionally, it assumes that each pseudo potential will always be
defined for a particular element which is why defining it is required
and it is exposed through a property. Since the base class cannot know
how the element can be automatically defined, if at all, it is just
checked in the `store` method that a valid element has been set.
Checking that it is commensurate with the actual file has to be left up
to the sub class implementation.