Add the `PseudoPotentialFamily` base class for pseudo potential families
The `PseudoPotentialFamily` is a `Group` entry point and serves as a
base class to define pseudo potential families. Subclasses just have to
define the type of pseudo potentials they expect, which has to be a
subclass of the `PseudoPotentialData` data plugin. This requirement
makes it possible to provide quite some functionality in this base
class, such as the creation of the family from a directory containing
the pseudo potential files, as well as various useful properties and
methods to get information about the family itself.