Add various utilities for the command line interface
The `attempt` context manager can be used in CLI commands to wrap a
function or method that can except. The manager will catch the exception
and print a formatted message instead.

The `create_family_from_archive` function simplifies constructing a
pseudo potential family form an archive file containing the pseudo
potential files. The subclass of `PseudoPotentialFamily` can be
specified.