Add the `SsspFamily` group plugin
This group is designed to represent SSSP configurations which are pseudo
potential families of carefully selected UPF pseudo potentials. Due to
the similarity, it extends the `UpfFamily` class but it puts additional
constraints on the labels. It should also restricit the exact pseudo
potential files to correspond only to those that are officially marked
by the SSSP. This requirement will have to be guaranteed by the CLI
command by only allowing installing it from archives downloaded from the
official SSSP resource.