Add the `aiida-pseudo install sssp` command
This command is different from the `aiida-pseudo install family` variant
because this one only allows installing the family from the official
SSSP repository, which is the Materials Cloud Archive entry. This is the
only way that we can guarantee that all `SsspFamily` instances actually
correspond to official configurations. Of course this can be
circumvented by instantiating an `SsspFamily` through the ORM, but that
is unavoidable.