`PseudoPotentialFamily.create_from_folder`: support nested directory
So far, the `create_from_folder` method only supported directories that
contained just the pseudo potentials and nothing else. However, often,
especially when unpacking an archive, the pseudos will be contained in a
subdirectory. Adding support for this is relatively trivial but will
make a huge difference in user experience.