`aiida-pseudo install family`: add support for URLs
So far, one could only install a pseudo family from an archive on the
local file system, but often they are made available on the web. The
interface now also accepts URLs and will automatically download the
resource locally. Since the command will receive a filelike-object but
the underlying method to perform the unpacking `shutil.unpack_archive`
does not accept those, the content first has to be written to a
temporary file on disk.

The `unpack_archive` method will automatically try to determine the
format of the archive from the filename extension so the temporary file
maintains that of the original file passed to the command. This might
still be incorrect, however, in which case the unpacking fails. In such
situations, users can use the new `-F/--archive-format` option to
define the format explicitly. The supported formats are all those that
are supported by `shutil` as returned by `get_archive_formats`.