Add `get_pseudos` to the `PseudoPotentialFamily` class
This method returns a dictionary that maps elements to the corresponding
pseudo potential data node that is stored in the family. The elements
can either be passed in directly as a list or tuple, or a `StructureData`
node can be used and the set of elements will be determined
automatically.