`UpfFamily`: add functionality to define recommended cutoffs
Codes that use pseudo potentials with the UPF format often also are
plane-wave codes which require the definition of energy cutoffs for the
wave functions and charge densities. Pseudo potential families can
provide recommended cutoffs for each pseudo potential making it easy to
determine which cutoffs should be used for a given set of elements.

The cutoffs are stored in the extras of the family and can be set with
the method `set_cutoffs`. The `get_cutoffs` method will return them.

This feature requires `aiida-core~=1.4` because that is the version that
introduced extras for groups, which is what is used to store the
recommended cutoffs for the pseudo potentials of a given family.