`UpfFamily`: add the `get_recommended_cutoffs` method
The `get_recommended_cutoffs` method allows to retrieve the recommended
cutoffs to use for the family based on a given set of elements. The
elements can be specified directly as a list or tuple, but there is also
support to pass a `StructureData` instance, in which case the elements
will be automatically determined. The function returns a tuple with the
highest wavefunction and charge density cutoff that are defined for the
set of provided elements.