`PseudoPotentialFamily`: deduplicate existing pseudos (#5)
When creating a new family through the `create_from_folder` class
method, the code will now check for each parsed pseudo potential whether
the exact same potential already exists in the database, and use that
instead. The check is performed on the exact pseudo potential node type
and the md5 of its file content. The behavior can be turned off through
the new `deduplicate` keyword.

The advantage of deduplication is not only a question of efficiency,
where it prevents that the same pseudo potential gets stored as a node
multiple times in the database, but it can also be more intuitive. If
one imports two families that each happen to contain the exact same
pseudo potential, one might expect that there is a single node that
represents that pseudo that just happens to be part of both groups that
represent the two pseudo potential families.