Pre-commit: replace old format string interpolation with f-strings
Since Python 3.5 is no longer supported, format string interpolations
can now be replaced by f-strings, introduced in Python 3.6, which are
more readable, require less characters and are more efficient.

All of the conversions were done automatically with the linting tool
`flynt` which is also added as a pre-commit hook. It is added before the
`yapf` step because since `flynt` will touch formatting, `yapf` will
then get a chance to check it.