`UpfFamily`: factor out cutoff functionality to `RecommendedCutoffMixin`
By factoring out the recommended cutoff functionality into a mixin
class, it can easily be added to whatever pseudopotential family that
also wants to support this feature.

To test the functionality, the mixin actually needs to be mixed into a
pseudopotential family class. A dummy class `CutoffFamily` is defined
for this purpose. However, this needs to update the metaclass that the
`Group` baseclass gets from `aiida-core`, because it won't be able to
define the `_type_string`, given that it does not have an associated
entry point. The `_type_string` being `None` would cause any query for
instances of the test class to except.