`RecommendedCutoffMixin`: add concept of stringency levels
Up till now, the `RecommendedCutoffMixin` allowed to define one single
set of recommended cutoffs per family. However, sometimes it makes sense
to define multiple sets of cutoffs for a single family, each that
represent a different level of computational stringency with respect to
the precision of the results that they produce.

The interface is updated to allow defining multiple sets of cutoffs. At
the same time a default stringency has to be defined such that the
normal interface `get_cutoffs` can continue to function without a caller
having to know which stringency to pass. The `get_cutoff_stringencies`
allows a caller to inspect which stringencies are available on the
family.