`PseudoPotentialFamily.get_pseudos`: maintain structure kind names (#20)
The method was incorrectly returning a mapping of the pseudos onto the
symbols of the structure's kind, thereby discarding the kind name, which
can often be different from the symbol. For example, `aiida-quantumespresso`
uses kind names that suffix the element symbol with a digit in order to
specify different magnetic kinds.