Remove pseudo potential family plugins for specific file formats
The initial design was to create a single pseudo potential family
subclass with corresponding entry point for each pseudo potential data
subtype available. Not only does this approach not really make sense, as
the format of the pseudo potential should be irrelevant to its content,
and so a family of UPF potentials can convey the exact same information
as one where the pseudos are formatted in PSP8 format, but it also
forces lots of subclasses and entry points to be created.

Still, having the format of the pseudos used in a pseudo family is still
useful, for example in order to query for it. Codes typically only
support a subset of pseudo formats, and so it is important to find a
family with the correct format when launching a calculation. The idea is
then to move this information to the extras of family such that it can
be queried for.

Here we first remove all existing pseudo family subclasses that only
exist because of a particular pseudo format.