`PseudoPotentialFamily`: override `remove_nodes` and `clear` (#29)
These methods need to be overridden, such that after they have called
the implementation of the `Group` base class, the internal `_pseudos`
class attribute can be updated accordingly. For `clear` this is easy as
it is simply set to `None`, but for `remove_nodes` special care needs to
be taken. It needs to account for the fact that a single node or an
iterable of nodes can be passed as an argument, and that the base class
implementation will not raise if any of the nodes is not actually
contained in the group. So instead of doing the straight forward thing
of removing the nodes from `_pseudos` based on the elements, we remove
them based on the pk, which guarantees we are removing the correct nodes.