`PseudoPotentialData`: add the `get_entry_point_name` classmethod (#30)
The method will return the entry point name under which the subclass of
`PseudoPotentialData` is registered. Note that the entry point name does
not include the entry point group. For example, the `UpfData` plugin is
registered under `aiida.data:pseudo.upf`, but the classmethod will
return the string `pseudo.upf`.