`PseudoPotentialData`: add the `get_or_create` classmethod
The classmethod has the exact same signature as the constructor but
instead of constructing a new instance straightaway, it will first check
if the database already contains a node with the exact same class (so no
subclassing is used in the query) and with an md5 checksum that matches
that of the bytes of the stream.

If an existing node is matched, that is returned, otherwise a new
instance is constructed and returned unstored.