`RecommendedCutoffMixin`: require electronvolts for cutoffs (#35)
Since multiple codes might start to use the `RecommendedCutoffMixin`
which might each require the cutoffs in different units, it is important
to have a standard unit. For now, electronvolt is chosen as the one and
only supported unit. Reason being that electronvolt is the default
energy unit for most of the AiiDA ecosphere up till now. In the future,
the mixin may be updated to support multiple units with automatic
conversion but this can be added later in a backwards compatible manner.