Add the `PseudoDojoFamily`
The Pseudo Dojo project provides multiple pseudopotential family
configurations for the functionals LDA, PBE and PBEsol. In addition, it
also provides families with various treatments for relativistic effects
and the pseudopotentials come in various formats, including PSP8, PSML,
UPF and JthXML. Just like the SSSP project, Pseudo Dojo provides
recommended cutoffs for plane waves and the charge density.

A new pseudopotential family class `PseudoDojoFamily` is added that
implements the logic to define valid configurations and how to retrieve
the pseudopotentials and recommended cutoffs that come with it.

Currently, certain configurations that are in principle valid, as in
they correspond to an existing family, are commented out. This is
because they are currently inconsistent. Either cutoffs are missing or
the md5 listed in the metadata does not match that of the included
pseudopotential files. Once these problems are addressed on the Pseudo
Dojo source, they can be uncommented such that they become available for
installation through this plugin.

Any family can be installed through `aiida-pseudo install pseudo-dojo`
which exposes options to select the desired configuration.