PseudoDojo: enable PAW based families and lower dual factors (#42)
The PseuDojo pseudopotential families with PAW bases pseudopotentials
are enabled for installation through the CLI. These configurations have
missing cutoffs for the Lanthanides, which have ben replaced with a
placeholder value of `-1` in the source metadata. The CLI replaces these
with the one and a half times the maximum cutoff found from the same
stringency set so that these potentials are still usable. However, this
should be taken as a _rough_ approximation.

In addition, the dual factors to compute the density cutoff from the
recommended wave function cutoff has been reduced by a factor of two for
the UPF, Psp8 and PSML families. This should still be plenty for
accurate results. The previous value was plain wasteful.