CLI: add recommended cutoffs to `aiida-pseudo show`
Previously, `aiida-pseudo show` merely shows the element, filename and
md5 of each pseudo contained within a family. Since the addition of the
`RecommendedCutoffMixin`, families can also specify recommended cutoffs.
Here we add these to the output of the command if the specified family
is a subclass of `RecommendedCutoffMixin`. If the family specifies
multiple stringencies, the `-s/--stringency` option can be used to
change the default stringency from being shown.

In order to test this, a generic pseudo potential family class with the
`RecommendedCutoffMixin` is needed. This was already implemented for the
tests of the mixin itself, which created a new group class on the fly
using the `aiida.orm.groups.GroupMeta` meta class to automatically set
the required type strings. This worked for these tests since they didn't
have to be queried for, however, for the `aiida-pseudo show` tests, the
CLI command will have to query for it and since the mock class does not
have an actual physical entry point, the query fails.

Since dynamically adding an entry point in the setup of the test suite
is not trivial, instead we add an actual `CutoffsFamily` class including
an entry point. This class is designed therefore to be mostly used for
testing purposes.