`RecommendedCutoffMixin`: store the unit of cutoffs in extras
By storing the unit of the cutoffs values explicitly in the extras, it
not only will give more certainty to the user of what the cutoffs mean
because they can inspect the unit, but the `get_recommended_cutoffs`
method can also provide the `unit` argument to automatically convert it
to another unit of energy if so desired.

The conversion between units is provided by the `pint` library, which is
added as a dependency. Users can specify units in the API through
strings using the names that are recognized by `pint`. Both the
shorthand, e.g. `eV`, as well as the long version `electron_volt` can be
used. Other common energy units are `Ry` and `Eh`, for Rydberg and
Hartree atomic units respectively.

The default unit that is set when defining recommended cutoffs is still
set to be electron volt, which was the implicitly assumed unit uptil now.
Families that have been created with versions preceding this commit,
will not actually have the unit defined in the unit, but
`get_cutoffs_unit` will return the default electron volt if it is
missing. This does mean that the `RecommendedCutoffMixin`DEFAULT_UNIT`
cannot be changed or this would lead to old families to be interpreted
to have the incorrect unit.