`PseudoDojoFamily`: skip repo files with unsupported extension (#48)
Some of the djrepo archives from the Pseudo Dojo contain extraneous
files with incorrect extensions (e.g. a second file for Ge with
extension `.djrepo_old`). This currently means that installing from this
archive raises a `ValueError` since duplicate elements will be found.

Here we add a check that makes sure the extension of the archive
filename is in a list of supported extensions, currently limited to
`.djrepo`. If the extension is incorrect, a warning is raised and the
file is skipped.