Docs: add troubleshooting section to the README.md (#60)
Add first entry with suggestion on how to install a pseudo potential
family manually with the CLI if the automated commands fail, which is
often possible for unstable internet connections.

Add disclaimer that the method currently does not install the
recommended cutoffs as the automated commands do, which will cause the
manually installed family to be incompatible with certain
functionalities of certain plugins that rely on those cutoffs. The
recently added command `aiida-pseudo family cutoffs set` allows to
define these manually as well.