CLI: Add `units` option to `aiida-pseudo family cutoffs set`
Currently the user is not able to specify the units of the plane-wave
energy cutoffs when using the `aiida-pseudo family cutoffs set` command.
This means that the user is limited to only specifying the units in
electronvolt.

Here we add the `-u/--unit` option to the `aiida-pseudo family cutoffs
set` command to allow to user to specify the units of the cutoff values
provided in the CUTOFFS JSON file.