CLI: do not override existing stringencies in `family cutoffs set` (#71)
The command's interface only allows to specify a single stringency at a
time, however, it would use the `set_cutoffs` method internally, which
would override any existing stringencies. This bug is addressed by first
retrieving existing stringencies before adding the new one to it and
setting the ensemble as the new stringencies.

There is an additional limitation that needs to be taken into account,
however. Since the current design only allows for a single energy unit
to be used at a time for all stringencies, when adding a new stringency,
the same unit has to be used. If, in the future, `set_cutoffs` starts to
support setting individual stringencies additively and the family
supports multiple units for the different stringencies, then all of this
logic can be simplified significantly and the limitations removed.