CLI: refactor download code from automated install commands (#69)
This serves two purposes: it makes it easier to reuse the download code
outside of the CLI and secondly, it enables the mocking of this part of
the code, which is a heavy operation. Without this, unit testing of this
command would be prohibitively expensive as it would need to download
all files for every test.

A monkeypatched version of the download functions is provided for both
the SSSP and PseudoDojo. A single test is added for each that simply
changes which configuration is installed.