CLI: add `--download-only` option to automatic family install commands (#65)
The `aiida-pseudo install sssp` and `aiida-pseudo install pseudo-dojo`
automatically download the required files for installing the
corresponding pseudopotential families. Since the entire operation can
sometimes fail, `aiida-pseudo install family` was added in order to
facilitate the manual installation without the automatic download.
Conversely, it may be useful to have the command only download the files
of a particular family, without actually installing them.

The `--download-only` option is added to both automatic family install
commands which will simply write the files to the current working
directory after having downloaded them. If the target file already
exists in the local working directory, it is not overwritten but a
warning is printed to say the file is skipped. The user can then simply
remove the clashing file and rerun the command.