`RecommendedCutoffMixin`: Add support for v0.5.0 pseudo families
Currently, pseudo families installed with `aiida-pseudo==0.5.0` will
fail to work with the new API features in to ways, both related to the
addition of units for the recommended cutoffs:

* `delete_cutoffs` expects the units to be stored in the
`'_cutoffs_unit'` extras dictionary. Simply adding `None` as a default
to return in the `.pop()` call fixes this issue.
* `get_cutoffs_unit` will return a `KeyError` since the previously
configured stringencies did not include units. Here we can add an `if`
clause that check if the requested stringency _does_ have its cutoffs
set. Since the default unit was and is `eV`, we set this as the new unit
for the stringency. From then on the pseudo family will be "fixed".