CLI: allow folder on disk for `aiida-pseudo install family` (#80)
Uptil now, the `install family` command only accepted (compressed)
archives of a directory containing the pseudopotential files. Often,
however, a user may already have an unpacked directory on local disk and
it would be ridiculous to force them to archive it first before being
able to use it with the CLI.

The parameter type of the `archive` argument is changed to use the
`PathOrUrl` type of `aiida-core`. This type will treat the value as a
URL and fetch its content if the value is not a valid filepath on the
local file system. This type is not ideal since its error handling is
different from the rest of the `aiida-pseudo` CLI so its a bit
inhomogeneous, but this will be improved later.

Additionally, the `--pseudo-type` option was added. Historically, the
command only exposed the `--family-type` option since in the beginning
of `aiida-pseudo` each family type defined its own hardcoded pseudo type
so it was implied. Since then, families can define their pseudo type per
instance and so now the `--pseudo-type` option has become necessary to
let the user specify which pseudopotential data class should be used.