Rename `CutoffsFamily` to `CutoffsPseudoPotentialFamily`
The `CutoffsFamily` was introduced in order to properly test the
`RecommendedCutoffsMixin`. It was even given an entry point but this was
just a requirement for the tests to work and it was never intended to
actually be used in production. However, there now seems an actual
usecase for this class as a substitute for the `SsspFamily` and
`PseudoDojoFamily` who also have the mixin and allow recommended cutoffs
to be defined but they come with strict rules and only official sets of
pseudopotentials are supported.

If one wants to create a custom pseudopotential family but still want to
use the recommended cutoff functionality, the `PseudoPotentialFamily`
base class doesn't suffice but the `CutoffsFamily` does. Now that we
actually consider the class to be used for production, a more aptly
named class name is required to properly reflect its intended usage.