CLI: block `family cutoffs set` for established families
One of the motivations of having dedicated and automated CLI commands to
install established pseudpotential families like the SSSP or the
Pseudo-dojo normconserving pseudopotentials is to try and prevent that
the pseudopotentials or recommended cutoffs deviate from the official
ones. However, the `family cutoffs set` command currently allows the
user to change the recommended cutoffs easily through the CLI.

In order to prevent this, we block usage of the command for `SsspFamily`
and `PseudoDojoFamily`'s, by adding an `exclude` input argument to the
`PseudoPotentialFamilyParam`, where the entry points of the classes that
should be disallowed can be provided. This is then used to adapt the
option decorator for the `PSEUDO_POTENTIAL_FAMILY` input argument of
`family cutoffs set`.

Note that the user _can_ still install e.g. the SSSP with their own
recommended cutoffs using the `install family` command as a
`CutoffsPseudoPotentialFamily`. However, the `SsspFamily` class is
reserved for pseudopotentials installed with the automated install
command.