CLI: add install `--from-dir` for sssp and pseudo-dojo
Our workaround for installing an "established" family like a `SsspFamily`
or `PseudoDojoFamily` required using the `install family` command, but
this is explicitly forbidden since we do not want the user to be able to
install an arbitrary set of pseudopotentials as one of these classes.
This means there is currently no way to install the archive and metadata
obtained with the `--download` option of the automated install commands.

Here, we add the `--from-dir` option to the automated install commands
to allow the user to install the downloaded pseudopotentials and set the
recommended cutoffs automatically for an `SsspFamily` or
`PseudoDojoFamily`.