`PseudoPotentialData`: allow `str` or `Path` for `source` argument (#98)
Currently, the source of the pseudopotential passed at construction time
of the `PseudoPotentialData` that represents it had to be a binary
stream. Here we introduce the possibility to allow the caller that pass
a `str` or `Path` object with the absolute path to the file content on
disk that should be wrapped. Note that, as opposed to the base class
`SinglefileData`, the filepath is not required to be absolute. In
addition, it raises a `FileNotFoundError` instead of a `ValueError` if
the provided filepath does not exist.

There is a potential inefficiency in the implementation after this change.
It is not only the constructor of `SinglefileData`, and therefore our
plugin `PseudoPotentialData` and all its subclasses, that needs to be
able to support the different source types and convert the paths to a
stream if that is what passed, the `set_file` method, which is called by
the constructor needs to apply the same checks and conversions. Here, we
need to therefore make sure that the conversion from filepath to byte
stream is only done once and not in the `set_file` call of every
subclass.

Therefore it is important that the `set_file` method first calls the
`prepare_source` class method, which will only convert to bytestream if
isn't already one, and that should then be passed to the superclass
call. Note that this should be called straight after the `prepare_source`
since the base class will set the actual file and filename attributes
that are methods that may be invoked afterwards will rely on.