Dependencies: put upper limit on `psycopg2-binary`
The release `psycopg2-binary==2.9` breaks the unit test manager of
`aiida-core`. This was fixed for in v1.6, however, that version no
longer supports Python 3.6, which falls back on `aiida-core==1.5.2` and
that doesn't apply the upper limit on `psycopg2-binary`, so we are
forced to do it here. This limit can be removed once we drop support for
Python 3.6.

Also fixes two minor things brought up by `pylint`.