CLI: fix bug in `install family` when downloading from URL
When the command was passed an URL it writes the retrieved content to a
temporary file on the local file system. Since the underlying
unarchiving code attempts to determine the format from the filename, it
was attempting to name the temporary file based on the the URL. This
worked for the old URL, since there (as luck would have it) the URL
ended with the filename containing the correct archive format extension.

However, this is clearly not always the case and it is in general
impossible to deduce the filename from the URL. It might not even contain
an actual filename. There is no other solution then have the user specify
the archive format with the `-f` option. It is not made required because
for the local files it might not be necessary.