Tests: do not rely on actual URL to test install family
The `install family` command allows to install a family from an archive
that is directly provided via an HTTP URL. This was being tested
explicitly by using a URL from the Materials Cloud Archive. However,
this approach of actually retrieving from a URL has two main downsides:

 * It will be potentially slow by having to download data from the web
 * It is susceptible to random failures: if the URL is (temporarily) not
   availabe, goes permanently goes offline, or the content changes.

That is why we  monkeypatch the ``PathOrUrl`` instead to simply return
the content of an archive that is created on the fly on the local disk.
The command expects that the parameter type returns an object that
contains the content of the archive under the ``content`` attribute.
To mimick this, we construct a namedtuple with a ``content`` attribute
on the fly.