Dependencies: update requirement `aiida-core~=2.0` (#120)
Adds compatibility with AiiDA v2.0. Notable changes:

* Entry points are now loaded through built-in modules instead of the
  custom `reentry`, so `reentry scan` no longer needs to be called.
* Tab-completion now ships with `click` itself, so `click_completion`
  and its manual activiation is removed.
* Update entry points from `aiida-core` by prefixing with `core.`

Also enable the `AIIDA_WARN_v3` environment variable in the `tests` job
of the CI workflow. This will print deprecation warnings from
`aiida-core`. These are also addressed in this commit:

* `Entity.objects` -> `Entity.collection`
* `Node.get_attribute` -> `Node.base.attributes.get`
* `Node.set_attribute` -> `Node.base.attributes.set`
* `Node.get_extra` -> `Node.base.extras.get`
* `Node.set_extra` -> `Node.base.extras.set`
* `Node.get_object_content` -> `Node.base.repository.get_object_content`
* `Node.add_incoming` -> `Node.base.links.add_incoming`

Due to the updated requirements of `click`, the `sphinx-click`
dependency of the `docs` extra also had to be upgraded because the lower
versions had an upper limit on `click`.