CLI: Reduce the load time significantly
Some of the commands imported a number of resources at the top of the
modules that have large loading times on first import. This was causing
the load time of the CLI to be ~1.5 seconds, even when just rendering
the help message.

By moving the slow packages inside the commands themselves, the load
time is reduced to a mere ~0.2 seconds. This makes the CLI much more
snappy to use, especially when using tab-completion.