`RecommendedCutoffMixin`: Remove workaround for stringency units
A workaround was added in `v0.6` for families with recommended cutoffs
that were installed with `v0.5` since that older version didn't
explicitly store the units of the stringencies in the extras which the
new API assumed would always be there.

The workaround would account for the lack of this extra and would
automatically introduce it (in a sense migrating the old family to be
compatible with the new API) the first time it would be requested.

Almost two years has passed since the release of `v0.6` which should
have been plenty for existing families to have been migrated through
this workaround and so it can now be removed.