DevOps: Update Python version and `setup-python` action in CI/CD
The `setup-python` action is updated to `v4` which comes with built-in
support for caching of the `pip` install step, so the `cache` action can
be dropped.

The default version of Python for non-matrix based jobs is updated to
3.10 which allows to use the more robust `removeprefix` string method
in the `validate_release_tag.py` file. This is more readable than the
previous string slicing.