CLI: Change `Critical` to `Report` if family exists in `install` command
The `aiida-pseudo install` command was printing a `CRITICAL` log message
if the requested family already exists. Since the desired goal is
achieved and so in reality the command didn't really fail, it makes more
sense to use a `REPORT` message instead. The command still returns a
non-zero exit code such that scripts can determine whether a new family
was actually installed or not.