CLI: Enable install `--download-only` without available profile
The `aiida-pseudo install sssp/pseudo-dojo` provide the `--download-only`
option to only download the archive. This is useful in case a family
needs to be installed on a machine without a (stable) internet
connection. With this option, the archive can be downloaded on another
machine first before transferring the file and installing it directly
from the archive.

Currently, however, the install command always requires an AiiDA profile
to be configured, even for the `--download-only` option in which case
the command never needs access to the profile. Here, this requirement is
removed by removing the `with_dbenv` decorator from the command.
Instead, the profile is loaded manually inside the command at the point
where it is really necessary.